feat: refactor configuration handling to use AppConfig and add shell execution preferences
