+++ crates/forge_api/src/forge_api.rs
@@ -164,10 +164,13 @@ impl<A: Services, F: CommandInfra> API for ForgeAPI<A, F> {
         self.services.read_app_config().await
     }
 
-    async fn modify_config<Fn>(&self, _f: Fn) -> Result<()>
+    async fn modify_config<Fx>(&self, f: Fx) -> Result<()>
     where
-        Fn: FnOnce(&mut ForgeConfig) + Send + Sync,
+        Fx: FnOnce(&mut AppConfig) + Send + Sync,
     {
-        todo!()
+        let mut config = self.services.read_app_config().await?;
+        f(&mut config);
+        self.services.write_app_config(&config).await?;
+        Ok(())
     }
 }
