+++ crates/forge_api/src/lib.rs
@@ -3,5 +3,5 @@ mod forge_api;
 
 pub use api::*;
 pub use forge_api::*;
-pub use forge_app::LoginInfo;
+pub use forge_app::{LoginInfo, AppConfig, ChoiceType, ConfigOption};
 pub use forge_domain::*;
