+++ crates/forge_app/src/app_config.rs
@@ -1,5 +1,6 @@
-use derive_more::From;
+use derive_more::{Display, From};
 use serde::{Deserialize, Serialize};
+use strum_macros::EnumIter;
 
 #[derive(Deserialize)]
 #[serde(rename_all = "camelCase")]
@@ -12,7 +13,10 @@ pub struct InitAuth {
 #[derive(Default, Clone, Serialize, Deserialize)]
 #[serde(rename_all = "camelCase")]
 pub struct AppConfig {
+    #[serde(default, skip_serializing_if = "Option::is_none")]
     pub key_info: Option<LoginInfo>,
+    #[serde(default, skip_serializing_if = "Option::is_none")]
+    pub choices: Option<ChoicesConfig>,
 }
 
 #[derive(Clone, Serialize, Deserialize, From)]
@@ -25,4 +29,26 @@ pub struct LoginInfo {
     pub email: Option<String>,
     #[serde(default, skip_serializing_if = "Option::is_none")]
     pub name: Option<String>,
+}
+
+#[derive(Default, Debug, Clone, Serialize, Deserialize, Display, PartialEq, EnumIter)]
+#[serde(rename_all = "camelCase")]
+pub enum ChoiceType {
+    Allow,
+    Reject,
+    #[default]
+    #[display("Ask every time")]
+    AskEveryTime,
+}
+
+#[derive(Default, Debug, Clone, Serialize, Deserialize)]
+#[serde(rename_all = "camelCase")]
+pub struct ChoicesConfig {
+    #[serde(default, skip_serializing_if = "Option::is_none")]
+    pub execute_shell_commands: Option<ChoiceType>,
+}
+#[derive(Clone, Debug, Display, EnumIter)]
+pub enum ConfigOption {
+    #[display("Shell Command Execution")]
+    ShellCommandExecution,
 }
