+++ crates/forge_domain/src/lib.rs
@@ -12,7 +12,6 @@ mod env;
 mod error;
 mod event;
 mod file;
-mod global_config;
 mod http_config;
 mod image;
 mod max_tokens;
@@ -60,7 +59,6 @@ pub use env::*;
 pub use error::*;
 pub use event::*;
 pub use file::*;
-pub use global_config::*;
 pub use http_config::*;
 pub use image::*;
 pub use max_tokens::*;
